pyproject.toml
[tool.poetry]
name = "document-merge-service"
version = "6.6.1"
description = "Merge Document Template Service"
license = "GPL-3.0-or-later"
authors = ["Adfinis AG <info@adfinis.com>"]
homepage = "https://github.com/adfinis/document-merge-service"
repository = "https://github.com/adfinis/document-merge-service"
documentation = "https://github.com/adfinis/document-merge-service/blob/main/README.md"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django :: 4",
    "Framework :: Django :: 4.2",
]
include = ["CHANGELOG.md"]
exclude = [
    "document-merge-service/**/tests",
]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
boto3 = "^1.34.143"
Babel = "^2.15.0"
Django = "~4.2.13"
django-cors-headers = "^4.4.0"
django-environ = "^0.11.2"
django-filter = "^24.2"
django-generic-api-permissions = "^0.4.6"
django-storages = "^1.14.3"
djangorestframework = "^3.15.2"
docx-mailmerge2 = "^0.8.0"
docxtpl = "^0.17.0"
Jinja2 = "^3.1.4"
mysqlclient = { version = "^2.2.4", optional = true }
pillow = "^10.3.0"
psycopg = { version = "^3.1.19", optional = true, extras = ["binary"] }
pymemcache = { version = "^4.0.0", optional = true }
python-dateutil = "^2.9.0"
python-memcached = "^1.59"
requests = "^2.32.3"
sentry-sdk = "^2.7.1"
urllib3 = "^2.2.1"
uWSGI = "^2.0.21"
xltpl = "~0.21"

[tool.poetry.group.dev.dependencies]
django-stubs = "5.0.2"
factory-boy = "3.3.0"
gitlint = "0.19.1"
mypy = "1.10.1"
pdbpp = "0.10.3"
psutil = "6.0.0"
pytest = "8.2.2"
pytest-cov = "5.0.0"
pytest-django = "4.8.0"
pytest-env = "1.1.3"
pytest-factoryboy = "2.7.0"
pytest-mock = "3.14.0"
pytest-randomly = "3.15.0"
python-semantic-release = "7.34.6"
requests-mock = "1.12.1"
syrupy = "4.6.1"
types-python-dateutil = "2.9.0.20240316"
types-requests = "2.32.0.20240712"
types-setuptools = "70.3.0.20240710"
types-toml = "0.10.8.20240310"
ruff = "^0.5.0"

[tool.poetry.extras]
mysql = ["mysqlclient"]
pgsql = ["psycopg"]
databases = ["mysqlclient", "psycopg"]
memcache = ["pymemcache"]
s3 = ["boto3"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    "migrations",
    "snapshots",
    ".venv",
]
line-length = 88

[tool.ruff.lint]
select = [
    "C9",
    "D",
    "E",
    "F",
    "W",
    "I",
]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D202",
    "E501",
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line
    "D407", # Missing dashed underline after section
    "D406", # Section name should end with a newline
    "D203", # one-blank-line-before-class (docstring)
]

[tool.ruff.lint.mccabe]
max-complexity = 11

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.pytest.ini_options]
addopts = "--reuse-db --randomly-seed=1521188766 --randomly-dont-reorganize"
DJANGO_SETTINGS_MODULE = "document_merge_service.settings"
filterwarnings = [
    "error::DeprecationWarning",
    "error::PendingDeprecationWarning",
    "ignore:invalid escape sequence",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
    "ignore:'imghdr' is deprecated:DeprecationWarning",
    "ignore:datetime.datetime.utcnow():DeprecationWarning"
]
env = [
    "ADMINS=Test Example <test@example.com>,Test2 <test2@example.com>",
    "OIDC_USERINFO_ENDPOINT=mock://document-merge-service.github.com/openid/userinfo",
    "OIDC_BEARER_TOKEN_REVALIDATION_TIME=60"
]

[tool.coverage.run]
source = ["."]

[tool.coverage.report]
fail_under = 100

exclude_lines = [
    "pragma: no cover",
    "pragma: todo cover",
    "def __str__",
    "def __unicode__",
    "def __repr__",
]
omit = [
    "*/migrations/*",
    "*/apps.py",
    "manage.py",
    "setup.py",
    "document_merge_service/settings_*.py",
    "document_merge_service/wsgi.py",
    "document_merge_service/document_merge_service_metadata.py",
]
show_missing = true

[tool.semantic_release]
version_source = "tag"
version_toml = ["pyproject.toml:tool.poetry.version"]
